Add predict method for avgSuperLearner class
Given the V different CATE super learners stored in an object of class avgSuperLearner, this method averages their predictions.

Code is also updated for when this method is called in compute_estimates.R. In this setting, we are predicting from the CATE models to get treatment assignments under the estimated treatment rule for observations in each outer validation fold.